APT: Rely on aptitude for "apt search"
/usr/bin/apt tries to do too much and doesn't do it well.

It brings info missing in "aptitude search" such as version and origins, but these are easily available via apt show and apt policy.

It lacks the following compared to aptitude:

 - It shows way too many results (doesn't just filter on pkg names)
 - It shows too much information and not in list format, which makes it hard to quickly see package names, or to grep/sort them
 - It doesn't support states or expressions (like for instance "aptitude search ~c")

It also breaks with the past. This command has been in Mint for years, long before /usr/bin/apt existed, so it needs to be better than before if it is to change.