Fix apt download
This feature was broken when switching apt to subprocess.

Also, the feature wasn't smart enough. It only really works for packages which are not installed.

In the case where the package is installed, switch to the inferior "/usr/bin/apt download",
which just downloads the deb for the given package.