Remove fortunes
This feature was fun but it had to be disabled by default because some fortune messages
were inapropriate, and some even insulting and racist.

New users didn't necessarily understand where the fortune message came from and
assumed it was authored by Linux Mint.

After a few years of this being disabled by default, and because there is no obvious
UI configuration tool to enable it, it's time to get rid of it entirely.