Bring back /usr/local/pastebin and make it use termbin
Mint 18.x used gist for pastebin (in mintsystem's /usr/local/bin in
18.0 and 18.1 and in xapps-common's /usr/bin in 18.2 and 18.3).

Gist no longer allows pastes. This is fixed in xapps-common in 19.x.

This commits brings back a /usr/local/bin/pastebin which uses termbin.
It will replace the file already there in 18.0 and 18.1 and override
the /usr/bin/pastebin in 18.2 and 18.3.